+++ db/unpatch_5.sql
@@ -1,0 +1,7 @@
+ALTER TABLE click
+DROP FOREIGN KEY click_impression_fk,
+DROP COLUMN impression_id;
+
+ALTER TABLE impression
+DROP INDEX impression_token,
+DROP COLUMN token;